requirements.txt
discord.py==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.3.0
python-dotenv==1.0.0
aiohttp==3.9.1
json5==0.9.14
markdown==3.5.1
pyyaml==6.0.1
schedule==1.2.0
nltk==3.8.1
textblob==0.17.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3 